Optimize zsh startup time
By looking at `zprof`'s output, it was clear that both `compdump` and
`compinit` were major culprits on the slow startup time (more than 1
second) that I was having.

This change improves that by adding a check on the `.zcompdump`'s
creation date and only recreating it if it is older than a day, by
removing a lot of superfluous `compinit` calls and by replacing or
removing all together some oh-my-zsh plugins.